+++ packages/headless/src/api/analytics/search-analytics.ts
@@ -4,15 +4,11 @@ import {
   history,
   CoveoAnalyticsClient,
   AnalyticsClientSendEventHook,
-  CaseAssistClient,
-  CaseAssistClientProvider,
 } from 'coveo.analytics';
 import {Logger} from 'pino';
 import {getQueryInitialState} from '../../features/query/query-state';
-import {getSearchHubInitialState} from '../../features/search-hub/search-hub-state';
 import {getSearchInitialState} from '../../features/search/search-state';
 import {
-  CaseAssistConfigurationSection,
   ConfigurationSection,
   ContextSection,
   FacetSection,
@@ -28,15 +24,15 @@ import {
   SearchSection,
 } from '../../state/state-sections';
 import {PreprocessRequest} from '../preprocess-request';
-import {getLanguage} from './shared-analytics';
+import {BaseAnalyticsProvider} from './base-analytics';
 import {
   buildFacetStateMetadata,
   SectionNeededForFacetMetadata,
   getStateNeededForFacetMetadata,
 } from '../../features/facets/facet-set/facet-set-analytics-actions-utils';
-import {VERSION} from '../../utils/version';
+import {SearchEventRequest} from 'coveo.analytics/dist/definitions/events';
 
-export type StateNeededByAnalyticsProvider = ConfigurationSection &
+export type StateNeededBySearchAnalyticsProvider = ConfigurationSection &
   Partial<
     SearchHubSection &
       SearchSection &
@@ -48,9 +44,13 @@ export type StateNeededByAnalyticsProvider = ConfigurationSection &
       SectionNeededForFacetMetadata
   >;
 
-export type StateNeededByCaseAssistAnalytics = ConfigurationSection &
-  Partial<CaseAssistConfigurationSection> &
-  Partial<SearchHubSection>;
+export class SearchAnalyticsProvider
+  extends BaseAnalyticsProvider
+  implements SearchPageClientProvider
+{
+  constructor(private state: StateNeededBySearchAnalyticsProvider) {
+    super(state);
+  }
 
 export type StateNeededByInsightAnalytics = ConfigurationSection &
   InsightConfigurationSection &
@@ -62,42 +62,8 @@ export type StateNeededByInsightAnalytics = ConfigurationSection &
       PaginationSection
   >;
 
-export class AnalyticsProvider implements SearchPageClientProvider {
-  constructor(private state: StateNeededByAnalyticsProvider) {}
-
-  public getLanguage() {
-    return getLanguage(this.state);
-  }
-
-  public getSearchEventRequestPayload() {
-    return {
-      queryText: this.queryText,
-      responseTime: this.responseTime,
-      results: this.mapResultsToAnalyticsDocument(),
-      numberOfResults: this.numberOfResults,
-    };
-  }
-
-  public getBaseMetadata() {
-    const {context} = this.state;
-    const contextValues = context?.contextValues || {};
-    const formattedObject: Record<string, string | string[]> = {};
-    for (const [key, value] of Object.entries(contextValues)) {
-      const formattedKey = `context_${key}`;
-      formattedObject[formattedKey] = value;
-    }
-    formattedObject['coveoHeadlessVersion'] = VERSION;
-    return formattedObject;
-  }
-
-  public getSearchUID() {
-    return (
-      this.state.search?.searchResponseId ||
-      this.state.search?.response.searchUid ||
-      this.state.recommendation?.searchUid ||
-      this.state.productListing?.responseId ||
-      getSearchInitialState().response.searchUid
-    );
+  public getFacetState() {
+    return buildFacetStateMetadata(getStateNeededForFacetMetadata(this.state));
   }
 
   public getPipeline() {
@@ -106,32 +72,24 @@ export class AnalyticsProvider implements SearchPageClientProvider {
     );
   }
 
-  public getOriginContext() {
-    return this.state.configuration.analytics.originContext;
+  public getSearchEventRequestPayload(): Omit<
+    SearchEventRequest,
+    'actionCause' | 'searchQueryUid'
+  > {
+    return {
+      queryText: this.queryText,
+      responseTime: this.responseTime,
+      results: this.mapResultsToAnalyticsDocument(),
+      numberOfResults: this.numberOfResults,
+    };
   }
 
-  public getOriginLevel1() {
-    return this.state.searchHub || getSearchHubInitialState();
-  }
-
-  public getOriginLevel2() {
-    return this.state.configuration.analytics.originLevel2;
-  }
-
-  public getOriginLevel3() {
-    // TODO: When referrer implemented;
-    // Configurable on headless engine, optionally
-    // If not specified at config time, need to fallback to use current referrer parameter for search API, if any
-    // Otherwise: fallback to `default`;
-    return this.state.configuration.analytics.originLevel3;
-  }
-
-  public getFacetState() {
-    return buildFacetStateMetadata(getStateNeededForFacetMetadata(this.state));
-  }
-
-  public getIsAnonymous() {
-    return this.state.configuration.analytics.anonymous;
+  public getSearchUID(): string {
+    return (
+      this.state.search?.searchResponseId ||
+      this.state.search?.response.searchUid ||
+      getSearchInitialState().response.searchUid
+    );
   }
 
   private mapResultsToAnalyticsDocument() {
@@ -157,16 +115,8 @@ export class AnalyticsProvider implements SearchPageClientProvider {
   }
 }
 
-export class CaseAssistAnalyticsProvider implements CaseAssistClientProvider {
-  constructor(private state: StateNeededByCaseAssistAnalytics) {}
-
-  public getOriginLevel1() {
-    return this.state.searchHub || getSearchHubInitialState();
-  }
-}
-
 interface ConfigureAnalyticsOptions {
-  state: StateNeededByAnalyticsProvider;
+  state: StateNeededBySearchAnalyticsProvider;
   logger: Logger;
   analyticsClientMiddleware?: AnalyticsClientSendEventHook;
   preprocessRequest?: PreprocessRequest;
@@ -178,7 +128,7 @@ export const configureAnalytics = ({
   state,
   analyticsClientMiddleware = (_, p) => p,
   preprocessRequest,
-  provider = new AnalyticsProvider(state),
+  provider = new SearchAnalyticsProvider(state),
 }: ConfigureAnalyticsOptions) => {
   const token = state.configuration.accessToken;
   const endpoint = state.configuration.analytics.apiBaseUrl;
@@ -231,55 +181,4 @@ export const getPageID = () => {
   }
 
   return lastPageView.value!;
-};
-
-interface ConfigureCaseAssistAnalyticsOptions {
-  state: StateNeededByCaseAssistAnalytics;
-  logger: Logger;
-  analyticsClientMiddleware?: AnalyticsClientSendEventHook;
-  preprocessRequest?: PreprocessRequest;
-  provider?: CaseAssistClientProvider;
-}
-
-export const configureCaseAssistAnalytics = ({
-  logger,
-  state,
-  analyticsClientMiddleware = (_, p) => p,
-  preprocessRequest,
-  provider = new CaseAssistAnalyticsProvider(state),
-}: ConfigureCaseAssistAnalyticsOptions) => {
-  const token = state.configuration.accessToken;
-  const endpoint = state.configuration.analytics.apiBaseUrl;
-  const runtimeEnvironment = state.configuration.analytics.runtimeEnvironment;
-  const enableAnalytics = state.configuration.analytics.enabled;
-  const client = new CaseAssistClient(
-    {
-      enableAnalytics,
-      token,
-      endpoint,
-      runtimeEnvironment,
-      preprocessRequest,
-      beforeSendHooks: [
-        analyticsClientMiddleware,
-        (type, payload) => {
-          logger.info(
-            {
-              ...payload,
-              type,
-              endpoint,
-              token,
-            },
-            'Analytics request'
-          );
-          return payload;
-        },
-      ],
-    },
-    provider
-  );
-
-  if (!enableAnalytics) {
-    client.disable();
-  }
-  return client;
 };